utility/requirements.txt
requests
requests_aws4auth
elasticsearch
boto3
PyGithub
